GCA-34 Introduce transaction management for vital actions - refactoring
