+++ src/main/java/com/gca/dao/transaction/Transactional.java
@@ -1,4 +1,4 @@
-package com.gca.security;
+package com.gca.dao.transaction;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -9,6 +9,6 @@ import java.lang.annotation.Target;
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-public @interface MyTransactional {
+public @interface Transactional {
     boolean readOnly() default false;
 }
