+++ src/main/java/com/gca/service/impl/TraineeServiceImpl.java
@@ -9,7 +9,7 @@ import com.gca.exception.ServiceException;
 import com.gca.mapper.TraineeMapper;
 import com.gca.model.Trainee;
 import com.gca.model.User;
-import com.gca.security.MyTransactional;
+import com.gca.dao.transaction.Transactional;
 import com.gca.service.TraineeService;
 import com.gca.service.common.CoreValidator;
 import jakarta.persistence.EntityNotFoundException;
@@ -55,7 +55,7 @@ public class TraineeServiceImpl implements TraineeService {
         this.validator = validator;
     }
 
-    @MyTransactional
+    @Transactional
     @Override
     public TraineeResponse createTrainee(@Valid TraineeCreateRequest request) {
         logger.debug("Creating trainee");
@@ -75,7 +75,7 @@ public class TraineeServiceImpl implements TraineeService {
         return traineeMapper.toResponse(created);
     }
 
-    @MyTransactional
+    @Transactional
     @Override
     public TraineeResponse updateTrainee(@Valid TraineeUpdateRequest request) {
         logger.debug("Updating trainee");
@@ -96,7 +96,7 @@ public class TraineeServiceImpl implements TraineeService {
         return traineeMapper.toResponse(updated);
     }
 
-    @MyTransactional(readOnly = true)
+    @Transactional(readOnly = true)
     @Override
     public TraineeResponse getTraineeById(Long id) {
         logger.debug("Retrieving trainee with ID: {}", id);
@@ -112,7 +112,7 @@ public class TraineeServiceImpl implements TraineeService {
         return traineeMapper.toResponse(trainee);
     }
 
-    @MyTransactional(readOnly = true)
+    @Transactional(readOnly = true)
     @Override
     public TraineeResponse getTraineeByUsername(String username) {
         logger.debug("Getting trainee by username: {}", username);
@@ -129,7 +129,7 @@ public class TraineeServiceImpl implements TraineeService {
                 ));
     }
 
-    @MyTransactional
+    @Transactional
     @Override
     public void deleteTraineeById(Long id) {
         logger.debug("Deleting trainee with ID: {}", id);
@@ -151,7 +151,7 @@ public class TraineeServiceImpl implements TraineeService {
                 );
     }
 
-    @MyTransactional
+    @Transactional
     @Override
     public void deleteTraineeByUsername(String username) {
         logger.debug("Deleting trainee by username: {}", username);
