+++ src/main/java/com/gca/service/impl/TrainerServiceImpl.java
@@ -11,7 +11,7 @@ import com.gca.mapper.TrainerMapper;
 import com.gca.model.Trainee;
 import com.gca.model.Trainer;
 import com.gca.model.User;
-import com.gca.security.MyTransactional;
+import com.gca.dao.transaction.Transactional;
 import com.gca.service.TrainerService;
 import com.gca.service.common.CoreValidator;
 import jakarta.persistence.EntityNotFoundException;
@@ -67,7 +67,7 @@ public class TrainerServiceImpl implements TrainerService {
         this.traineeDAO = traineeDAO;
     }
 
-    @MyTransactional
+    @Transactional
     @Override
     public TrainerResponse createTrainer(@Valid TrainerCreateRequest request) {
         logger.debug("Creating trainer");
@@ -87,7 +87,7 @@ public class TrainerServiceImpl implements TrainerService {
         return trainerMapper.toResponse(created);
     }
 
-    @MyTransactional
+    @Transactional
     @Override
     public TrainerResponse updateTrainer(@Valid TrainerUpdateRequest request) {
         logger.debug("Updating trainer");
@@ -108,7 +108,7 @@ public class TrainerServiceImpl implements TrainerService {
         return trainerMapper.toResponse(updated);
     }
 
-    @MyTransactional(readOnly = true)
+    @Transactional(readOnly = true)
     @Override
     public TrainerResponse getTrainerByUsername(String username) {
         logger.debug("Getting trainer by username: {}", username);
@@ -125,7 +125,7 @@ public class TrainerServiceImpl implements TrainerService {
                 ));
     }
 
-    @MyTransactional(readOnly = true)
+    @Transactional(readOnly = true)
     @Override
     public List<Trainer> getUnassignedTrainers(String traineeUsername) {
         logger.debug("Getting unassigned trainers for trainee username: {}", traineeUsername);
@@ -148,7 +148,7 @@ public class TrainerServiceImpl implements TrainerService {
         return unassignedTrainers;
     }
 
-    @MyTransactional(readOnly = true)
+    @Transactional(readOnly = true)
     @Override
     public TrainerResponse getTrainerById(Long id) {
         logger.debug("Retrieving trainer by ID: {}", id);
